+++ modules/spring-core/src/test/kotlin/com/explyt/spring/core/inspections/java/SpringYamlInspectionTest.kt
@@ -80,4 +80,100 @@ explyt.prop:
         )
         myFixture.testHighlighting("application.yaml")
     }
+
+    fun testFileDefinitionPropertiesSubstring() {
+        @Language("java") val configurationProperty = """
+            @org.springframework.context.annotation.Configuration
+            @org.springframework.boot.context.properties.ConfigurationProperties(prefix = "explyt-prop")
+            public class ConfigProperties {               
+                private String test;                           
+                public void setTest(String test) { this.test = test}
+                public String getTest() { return this.test}
+            } 
+        """.trimIndent()
+        myFixture.addClass(configurationProperty)
+        myFixture.configureByText(
+            "application.yaml",
+            """
+explyt.prop:
+    test: some1
+explyt: some2
+            """.trimIndent()
+        )
+        myFixture.testHighlighting("application.yaml")
+    }
+
+    fun testComplexKebabCaseProperties() {
+        myFixture.configureByText(
+            "application.yaml",
+            """
+foo:
+    <warning descr="Should be kebab-case">barbaz.testProp</warning>: some1    
+            """.trimIndent()
+        )
+        myFixture.testHighlighting("application.yaml")
+    }
+
+    fun testKebabCaseQuickFix() {
+        myFixture.configureByText(
+            "application.yaml",
+            """
+fooFoo:
+  bar:
+    baz: 5   
+            """.trimIndent()
+        )
+        val quickFix = myFixture.getAllQuickFixes().firstOrNull()
+        assertNotNull(quickFix)
+        myFixture.launchAction(quickFix!!)
+        myFixture.checkResult(
+            """
+foo-foo:
+  bar:
+    baz: 5 
+        """.trimIndent(), true
+        )
+    }
+
+    fun testKebabCaseQuickFixMiddle() {
+        myFixture.configureByText(
+            "application.yaml",
+            """
+foo:
+  barFoo:
+    baz: 5   
+            """.trimIndent()
+        )
+        val quickFix = myFixture.getAllQuickFixes().firstOrNull()
+        assertNotNull(quickFix)
+        myFixture.launchAction(quickFix!!)
+        myFixture.checkResult(
+            """
+foo:
+  bar-foo:
+    baz: 5 
+        """.trimIndent(), true
+        )
+    }
+
+    fun testKebabCaseQuickFixEnd() {
+        myFixture.configureByText(
+            "application.yaml",
+            """
+foo:
+  bar:
+    bazFoo: 5   
+            """.trimIndent()
+        )
+        val quickFix = myFixture.getAllQuickFixes().firstOrNull()
+        assertNotNull(quickFix)
+        myFixture.launchAction(quickFix!!)
+        myFixture.checkResult(
+            """
+foo:
+  bar:
+    baz-foo: 5 
+        """.trimIndent(), true
+        )
+    }
 }